cmd: handle graceful shutdowns
probably not much difference from a normal ctrl-c at this point, but
demonstrates best practices for someone using this within a subroutine
of their own program where the binary wont immediately exit and cleanup
matters.

also, if the Updater switches to using RTM long-lived sockets in the
future, this will enable the CLI to gracefully disconnect, etc.